type casting in proto to make work on AVR

+++ lib/c/proto.c
@@ -266,13 +266,13 @@ void proto_recv(struct proto *p, uint8_t data)
 
 	case PROTO_STATE_LONG_DATA2:
 		p->long_val &= 0x00ff0000;
-		p->long_val |= data << 16;
+		p->long_val |= (uint32_t)data << 16;
 		p->state = PROTO_STATE_LONG_DATA3;
 		break;
 
 	case PROTO_STATE_LONG_DATA3:
 		p->long_val &= 0xff000000;
-		p->long_val |= data << 24;
+		p->long_val |= (uint32_t)data << 24;
 
 		if (p->packet.addr == p->addr && p->handlers->long_data)
 			p->handlers->long_data(p->handler_data, p->long_val);
